third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/object-src-none-blocked-expected.txt

CONSOLE ERROR: Refused to load plugin data from 'data:application/x-blink-test-plugin,' because it violates the following Content Security Policy directive: "object-src 'none'".



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
